Enhanced windows portability
git-svn-id: svn://scm.gforge.inria.fr/svn/openvibe@68 c330d7e9-fc0c-0410-a5b3-fd85c6f5aa8f

+++ openvibe-modules/system/trunc/src/system/Memory.cpp
@@ -9,6 +9,7 @@ boolean Memory::copy(
 	const void* pSourceBuffer,
 	const uint64 ui64BufferSize)
 {
-	memcpy(pTargetBuffer, pSourceBuffer, ui64BufferSize);
+	// $$$ TODO take 64bits size into consideration
+	memcpy(pTargetBuffer, pSourceBuffer, static_cast<size_t>(ui64BufferSize));
 	return true;
 }
